requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
Faker==18.3.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
